consolidated all files into the react app
