--- .gitignore
@@ -1,19 +1,0 @@
-# Ignore virtual environment
-venv/
-
-# Byte-compiled files
-__pycache__/
-*.pyc
-*.pyo
-
-# Environment variables
-.env
-
-# VS Code settings (if you use VS Code)
-.vscode/
-
-# Ignore any logs
-*.log
-
-# Ignore MacOS system files
-.DS_Store
