+++ thurs_night/hardwareSet.py
